# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ChatAttributes": {
				"properties": {
					"title": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"id"
				],
				"type": "object"
			},
			"Error": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Required_ChatInput_": {
				"properties": {
					"title": {
						"type": "string"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"description": "Make all properties in T required"
			},
			"ChatCreateParams": {
				"$ref": "#/components/schemas/Required_ChatInput_"
			},
			"Required_ChatAttributes_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title"
				],
				"type": "object",
				"description": "Make all properties in T required"
			},
			"ChatOutput": {
				"$ref": "#/components/schemas/Required_ChatAttributes_"
			},
			"LINE_ROLE": {
				"enum": [
					"Human",
					"AI"
				],
				"type": "string"
			},
			"Required_LineAttributes_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"text": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/LINE_ROLE"
					}
				},
				"required": [
					"id",
					"text",
					"role"
				],
				"type": "object",
				"description": "Make all properties in T required"
			},
			"LineOutput": {
				"$ref": "#/components/schemas/Required_LineAttributes_"
			},
			"ChatDetailOutput": {
				"properties": {
					"lines": {
						"items": {
							"$ref": "#/components/schemas/LineOutput"
						},
						"type": "array"
					},
					"chat": {
						"$ref": "#/components/schemas/ChatOutput"
					}
				},
				"required": [
					"lines",
					"chat"
				],
				"type": "object"
			},
			"LineGenerateResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"history": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"text": {
						"type": "string"
					}
				},
				"required": [
					"history",
					"text"
				],
				"type": "object"
			},
			"LineGenerateParams": {
				"properties": {
					"text": {
						"type": "string"
					},
					"chatId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"text",
					"chatId"
				],
				"type": "object"
			},
			"ChatLinesResponse": {
				"properties": {
					"lines": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"lines"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "text-completion-api",
		"version": "0.0.1",
		"description": "A Text Completion API that predicts the next word in a given text.",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "chrisyao4700 ",
			"email": "chrisyao900908@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/chat": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Created Chat",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ChatAttributes"
										},
										{
											"$ref": "#/components/schemas/Error"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Body that is sent for create a Chat",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChatCreateParams",
								"description": "Body that is sent for create a Chat"
							},
							"example": {
								"title": "New Chat"
							}
						}
					}
				}
			}
		},
		"/v1/chat/detail/{chatId}": {
			"get": {
				"operationId": "Find",
				"responses": {
					"200": {
						"description": "Chat Found",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ChatDetailOutput"
										},
										{
											"$ref": "#/components/schemas/Error"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "chatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/line": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Created text",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/LineGenerateResponse"
										},
										{},
										{
											"$ref": "#/components/schemas/Error"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Body that is sent for create a Line",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LineGenerateParams",
								"description": "Body that is sent for create a Line"
							},
							"example": {
								"text": "123",
								"chatId": 12
							}
						}
					}
				}
			}
		},
		"/v1/line/all/{chatId}": {
			"get": {
				"operationId": "GetLines",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ChatLinesResponse"
										},
										{},
										{
											"$ref": "#/components/schemas/Error"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "chatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}